Move Process.Viewable to its own module
